% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ssgsea.R
\name{build_ssgsea}
\alias{build_ssgsea}
\title{Perform ssGSEA pipeline for related signatures in defined genome.}
\usage{
build_ssgsea(genome, signatures, ...)
}
\arguments{
\item{genome}{Genome expression object in matrix format, with samples in columns and genes in rows.}

\item{signatures}{Signatures list object}

\item{...}{GSVA ssgsea extra parameters}
}
\description{
Encapsulates ssGSEA pipeline at GSVA package
}
