% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_model_coefficients.R
\name{extract_model_coefficients}
\alias{extract_model_coefficients}
\title{extract_model_coefficients}
\usage{
extract_model_coefficients(model, main_covariable)
}
\arguments{
\item{model}{Model  Fitted model}

\item{main_covariable}{String  The main covariable label, this is the studied one, other covaraibles can be considered or not at the regression}
}
\value{
DataFrame  The coefficients of the main covariable with the p value and adjusted R squared of the fitted model
}
\description{
This function extracts the coefficients of the main covariable and returns them along the p value of the model into a data frame
}
