% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_effect_variable.R
\name{preprocess_effect_variable}
\alias{preprocess_effect_variable}
\title{Preprocess effect variable
Applies bestNormalize package (optionally) removes outliers (defined by 1.5 IQR)}
\usage{
preprocess_effect_variable(values, remove_outliers = FALSE, ...)
}
\arguments{
\item{values}{Array with values}

\item{remove_outliers}{Flag to detect potential outliers and "remove" them setting as NA.}

\item{...}{Extra parameters for bestNormalize function}
}
\value{
out Array with clean data
}
\description{
Preprocess effect variable
Applies bestNormalize package (optionally) removes outliers (defined by 1.5 IQR)
}
