% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_covar_table.R
\name{preprocess_phenotypes}
\alias{preprocess_phenotypes}
\title{Cleans a phenotypes dataframe, tlhis is a helper function for systematic analysis.}
\usage{
preprocess_phenotypes(phenotypes, samples_colname, phenotypes_to_include)
}
\arguments{
\item{phenotypes}{data_frame with at least a samples column and some phenotype columns to be included.}

\item{samples_colname}{string Column with the samples data.}

\item{phenotypes_to_include}{String list with the list of phenotypes which should be included.}
}
\value{
DataFrame Rownames are the sample ids (of sample_colnames) and content
   contains the included phenotypes.
}
\description{
Removes samples with at leas one NA value and phenotypes where all values describe the same values.
}
