% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Detects which values lies beyond the extremes of the whiskers of a box plot and marks them}
\usage{
remove_outliers(values)
}
\arguments{
\item{values}{Array  A numeric array to mark outliers}
}
\value{
A numeric array with outliers value as NA
}
\description{
Detects which values lies beyond the extremes of the whiskers of a box plot and marks them
}
