% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_fit_model.R
\name{stepwise_fit_model}
\alias{stepwise_fit_model}
\title{Fit best model linear regression model using AIC as criterion.}
\usage{
stepwise_fit_model(y, phenotypes, ...)
}
\arguments{
\item{y}{Array of values}

\item{phenotypes}{Data frame with phenotype-related covariables}

\item{...}{List of other parameters related with MASS::stepAIC function}
}
\value{
List with base and fitted linear regression models.
}
\description{
Two models are returned, a base model which includes all values at phenotypes dataframe, and
a stepwise fitted model using Akaike Information Criterion.
}
\examples{
\dontrun{
  stepwise_fit_model(
    y = rnorm(100, 1, 10),
    data.frame(covar = rnorm(100,  1, 1), covar2 = rnorm(100, 1, 1) )
  )


  stepwise_fit_model(
    y = rnorm(100, 1, 10),
    data.frame(covar = rnorm(100,  1, 1), covar2 = rnorm(100, 1, 1) ),
    scope = list(
      lower = ~ covar,
      upper = ~ .
    )
  )

}
}
