% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systematic_regression.R
\name{systematic_regression}
\alias{systematic_regression}
\title{Systematic regression function which performs the complete pipeline and returns the results flattened into a data frame row}
\usage{
systematic_regression(
  dataset,
  effect_variable,
  main_covariable,
  sample_variable = "sample",
  phenotype_variables = "ALL"
)
}
\arguments{
\item{dataset}{DataFrame This dataframe contains the whole data, combining sample ids, phenotypes and independent variables}

\item{effect_variable}{String The dependent variable label}

\item{main_covariable}{String The main covariable label, this is the studied one, other covaraibles can be considered or not at the regression.}

\item{sample_variable}{String The name of the column where the sample names are defined.}

\item{phenotype_variables}{Array|"ALL" The list of covariables that cuould be included. Columns not defined in this array will be excluded from the study.}
}
\value{
DataFrame           One row dataframe with all the phentoype values
}
\description{
Systematic regression function which performs the complete pipeline and returns the results flattened into a data frame row
}
